feat: inline css on release
