+++ Cargo.toml
@@ -11,6 +11,9 @@ license = "MIT OR Apache-2.0"
 [[bin]]
 name = "oranda"
 
+[workspace]
+members = ["oranda-css"]
+
 [dependencies]
 ammonia = "3"
 axoasset = { version = "0.4.0", features = ["json-serde", "toml-edit"] }
@@ -50,11 +53,14 @@ pathdiff = { version = "0.2.1", features = ["camino"] }
 minijinja = { version = "1.0.3", features = ["loader", "preserve_order"] }
 include_dir = "0.7.3"
 slug = "0.1.4"
-directories = "5.0.1"
+oranda-generate-css = { path = "oranda-css" }
 
 [dev-dependencies]
 assert_cmd = "2"
 assert_fs = "1.0.7"
+
+[build-dependencies]
+oranda-generate-css = { path = "oranda-css" }
 
 # The profile that 'cargo dist' will build with
 [profile.dist]
